git-svn-id: file:///srv/local/data/cleanRepo/trunk@30834 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/johannes/src/main/java/playground/johannes/gsv/synPop/data/LandUseDataLoader.java
@@ -73,7 +73,7 @@ public class LandUseDataLoader implements DataLoader {
 					Object value = zone.getAttribute().get(popKey);
 					if(value != null) {
 						double d = Double.parseDouble(value.toString());
-						zone.getAttribute().put(LandUseData.POPULATION_KEY, (int)d);
+						zone.getAttribute().put(LandUseData.POPULATION_KEY, d);
 					}
 				}
